151 add can activate guard service
If the user changes some configuration, and click back arrow, we need to ask the user if he or she want to discard the changes or not.

First we need to add a can activate guard service, this service create an interface called CanComponentDeactivate, this interface has a method: canDeactivate, the return value of this instance should be either observable, promise or boolean.

then in this service, we create a class and implement this interface. canDeactivate has four params: component, currentRoute, currentState, nextState.